2 implement base support for heat relays (#11)
* Added seed args.

* Adde shelly 2 suport

* Update 2

* Update 3

* First working

* Almost there

* Styling

+++ AHeat.Web.Client/Pages/Admin/Roles/Index.razor.cs
@@ -27,11 +27,11 @@ public partial class Index
         Roles = await RolesClient.GetRolesAsync();
     }
 
-    private async Task AddEditRole(RoleDto role)
+    private async Task AddEditRole(RoleDto role, bool add)
     {
         var parameters = new DialogParameters<AddEditRoleDialog> { { x => x.role, role } };
 
-        var dialog = await DialogService.ShowAsync<AddEditRoleDialog>("Add Role", parameters);
+        var dialog = await DialogService.ShowAsync<AddEditRoleDialog>(add ? "Add Role!" : "Edit Role!", parameters);
         var result = await dialog.Result;
 
         if (!result.Canceled)
@@ -61,7 +61,7 @@ public partial class Index
             "Warning",
             $"Delete role {role.Name} ?",
             yesText: "Delete!", cancelText: "Cancel");
-        if (result.Value) 
+        if (result != null && result.Value)
         {
             await RolesClient.DeleteRoleAsync(role.Id);
             Roles.Remove(role);
